backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-groq
langgraph
sqlalchemy
psycopg2-binary